requirements.txt
numpy>=1.20.0
pandas>=1.3.0
yfinance>=0.1.70
scipy>=1.7.0
matplotlib>=3.5.0
dash>=2.5.0
plotly>=5.8.0
torch>=1.11.0
scikit-learn>=1.0.0
ephem>=4.1.0
pyswisseph>=2.10.0